Information about role group mapping is saving during job saving(#1671)
