+++ src/main/java/org/ohdsi/webapi/user/importer/converter/BaseUserImportJobToUserImportJobDTOConverter.java
@@ -12,5 +12,7 @@ public abstract class BaseUserImportJobToUserImportJobDTOConverter<T extends Use
     target.setProviderType(source.getProviderType());
     target.setPreserveRoles(source.getPreserveRoles());
     target.setUserRoles(source.getUserRoles());
+    target.setRoleGroupMapping(RoleGroupMappingConverter.convertRoleGroupMapping(source.getProviderType().getValue(),
+            source.getRoleGroupMapping()));
   }
 }
