+++ src/main/java/org/ohdsi/webapi/user/importer/dto/UserImportJobDTO.java
@@ -2,6 +2,7 @@ package org.ohdsi.webapi.user.importer.dto;
 
 import com.odysseusinc.scheduler.api.v1.dto.ArachneJobDTO;
 import org.ohdsi.webapi.user.importer.model.LdapProviderType;
+import org.ohdsi.webapi.user.importer.model.RoleGroupMapping;
 
 import java.util.Date;
 
@@ -17,6 +18,16 @@ public class UserImportJobDTO extends ArachneJobDTO {
   private Date nextExecution;
 
   private Date startDate;
+
+  private RoleGroupMapping roleGroupMapping;
+
+  public RoleGroupMapping getRoleGroupMapping() {
+    return roleGroupMapping;
+  }
+
+  public void setRoleGroupMapping(RoleGroupMapping roleGroupMapping) {
+    this.roleGroupMapping = roleGroupMapping;
+  }
 
   public LdapProviderType getProviderType() {
     return providerType;
