--- src/main/java/org/ohdsi/webapi/user/importer/dto/UserImportJobMappingDTO.java
@@ -1,16 +1,0 @@
-package org.ohdsi.webapi.user.importer.dto;
-
-import org.ohdsi.webapi.user.importer.model.RoleGroupMapping;
-
-public class UserImportJobMappingDTO extends UserImportJobDTO {
-
-  private RoleGroupMapping roleGroupMapping;
-
-  public RoleGroupMapping getRoleGroupMapping() {
-    return roleGroupMapping;
-  }
-
-  public void setRoleGroupMapping(RoleGroupMapping roleGroupMapping) {
-    this.roleGroupMapping = roleGroupMapping;
-  }
-}
